LPS-54752 Reorganize implementations of Settings interface into their own package inside portal-service. This is to make the API cleaner and avoid mixing so many classes into one only package.
