+++ portal-impl/test/unit/com/liferay/portal/kernel/settings/TypedSettingsTest.java
@@ -14,6 +14,7 @@
 
 package com.liferay.portal.kernel.settings;
 
+import com.liferay.portal.kernel.settings.impl.MemorySettings;
 import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.LocalizationUtil;
 import com.liferay.portal.util.LocalizationImpl;
