+++ portal-service/src/com/liferay/portal/kernel/settings/SettingsLocatorHelper.java
@@ -18,6 +18,10 @@ import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.resource.manager.ClassLoaderResourceManager;
 import com.liferay.portal.kernel.resource.manager.ResourceManager;
 import com.liferay.portal.kernel.security.pacl.DoPrivileged;
+import com.liferay.portal.kernel.settings.impl.ConfigurationBeanSettings;
+import com.liferay.portal.kernel.settings.impl.LocationVariableResolver;
+import com.liferay.portal.kernel.settings.impl.PortletPreferencesSettings;
+import com.liferay.portal.kernel.settings.impl.PropertiesSettings;
 import com.liferay.portal.kernel.util.PortalClassLoaderUtil;
 import com.liferay.portal.kernel.util.PropsUtil;
 import com.liferay.portal.model.Group;
