+++ portal-service/src/com/liferay/portal/kernel/settings/impl/ConfigurationBeanSettings.java
@@ -12,9 +12,12 @@
  * details.
  */
 
-package com.liferay.portal.kernel.settings;
+package com.liferay.portal.kernel.settings.impl;
 
 import com.liferay.portal.kernel.exception.SystemException;
+import com.liferay.portal.kernel.settings.BaseSettings;
+import com.liferay.portal.kernel.settings.LocalizedValuesMap;
+import com.liferay.portal.kernel.settings.Settings;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
