+++ portal-service/src/com/liferay/portal/kernel/settings/impl/FallbackSettings.java
@@ -12,7 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.kernel.settings;
+package com.liferay.portal.kernel.settings.impl;
+
+import com.liferay.portal.kernel.settings.BaseSettings;
+import com.liferay.portal.kernel.settings.FallbackKeys;
+import com.liferay.portal.kernel.settings.Settings;
 
 /**
  * @author Iván Zaera
