+++ portal-service/test/unit/com/liferay/portal/kernel/settings/BaseSettingsTest.java
@@ -14,6 +14,9 @@
 
 package com.liferay.portal.kernel.settings;
 
+import com.liferay.portal.kernel.settings.impl.MemorySettings;
+import com.liferay.portal.kernel.settings.impl.ParameterMapSettings;
+
 import java.util.Collections;
 
 import org.junit.Assert;
