+++ portal-service/test/unit/com/liferay/portal/kernel/settings/impl/FallbackSettingsTest.java
@@ -12,7 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.kernel.settings;
+package com.liferay.portal.kernel.settings.impl;
+
+import com.liferay.portal.kernel.settings.FallbackKeys;
+import com.liferay.portal.kernel.settings.Settings;
 
 import org.junit.Assert;
 import org.junit.Test;
