+++ portal-service/test/unit/com/liferay/portal/kernel/settings/impl/PropertiesSettingsTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.settings;
+package com.liferay.portal.kernel.settings.impl;
 
 import java.util.Properties;
